requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8
mypy==1.7.1
django-stubs==4.2.7
types-redis==4.6.0.20240903
pytz==2023.3
psycopg[binary]==3.1.18 